kernelci.legacy: move push_tarball() to kernelci.legacy
Move the kernelci.build.push_tarball() function to kernelci.legacy as
the new pipeline has a separate implementation based on
kernelci.storage instead.  This code can be dropped when the new API
replaces the legacy kernelci-backend in production.
