+++ kernelci/legacy.py
@@ -15,7 +15,10 @@
 # along with this library; if not, write to the Free Software Foundation, Inc.,
 # 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA
 
-from kernelci.build import get_branch_head
+import os
+import requests
+from urllib.parse import urljoin
+from kernelci.build import get_branch_head, git_describe, make_tarball
 from kernelci.storage import upload_files
 
 
@@ -71,3 +74,33 @@ def check_new_commit(config, storage):
         return True
     else:
         return branch_head
+
+
+def push_tarball(config, kdir, storage, api, token):
+    """Create and push a linux kernel source tarball to the storage server
+
+    If a tarball with a same name is already on the storage server, no new
+    tarball is uploaded.  Otherwise, a tarball is created
+
+    *config* is a BuildConfig object
+    *kdir* is the path to a kernel source directory
+    *storage* is the base URL of the storage server
+    *api* is the URL of the KernelCI backend API
+    *token* is the token to use with the KernelCI backend API
+
+    The returned value is the URL of the uploaded tarball.
+    """
+    tarball_name = "linux-src_{}.tar.gz".format(config.name)
+    describe = git_describe(config.tree.name, kdir)
+    path = '/'.join(list(item.replace('/', '-') for item in [
+        config.tree.name, config.branch, describe
+    ]))
+    tarball_url = urljoin(storage, '/'.join([path, tarball_name]))
+    resp = requests.head(tarball_url)
+    if resp.status_code == 200:
+        return tarball_url
+    tarball = "{}.tar.gz".format(config.name)
+    make_tarball(kdir, tarball)
+    upload_files(api, token, path, {tarball_name: open(tarball, 'rb')})
+    os.unlink(tarball)
+    return tarball_url
